extracted debug implementation
